refactor:move CuratedWorkAttribute into model folder
