+++ src/app/page.tsx
@@ -2,7 +2,7 @@
 import { ArtworkCarousel } from '../components/home/artworkCarousel';
 import { ArtworkQuiz } from '../components/home/artworkQuiz';
 import { MCQAttribute } from '@/model/interface/MCQ';
-import { CuratedWorkAttribute } from '@/types/curatedArtwork-types';
+import { CuratedWorkAttribute } from '@/model/interface/curatedArtwork-types';
 
 // TODO 배포할때 더 좋은 방법이 있을지 생각해보기
 const SERVER_URL =
